% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with 15 rows and 2 columns.
}
\usage{
fire
}
\description{
Relates distance of a fire from the nearest fire station to its financial damage.
}
\keyword{datasets}
