% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{myboot_Ass4}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{a}

\item{x}{b}

\item{fun}{c}

\item{alpha}{d}

\item{...}{e}
}
\value{
f
}
\description{
myboot_Ass4
}
