% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Function -- Uniform}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{number of elements}

\item{iter}{number of iterations}

\item{a}{min}

\item{b}{max}
}
\value{
Histogram for central limit of given values
}
\description{
Central Limit Function -- Uniform
}
\examples{
mycltu(n=20,iter=100000)
}
