% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{log likelihood}

\item{x}{variable}

\item{param}{parameter to be measured}

\item{...}{the rest}
}
\value{
MLE using grid approximation
}
\description{
mymaxlik
}
