% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats in flight}

\item{gamma}{Probability a plane will be truly overbooked: shows > seats}

\item{p}{Probability of a show}
}
\value{
list and Objective plots (binom and norm)
}
\description{
ntickets
}
\examples{
ntickets(200, 0.02, 0.95)
}
