% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{mycltu}
\alias{mycltu}
\title{CI for mean of a sample}
\usage{
mycltu(d, alpha)
}
\arguments{
\item{d}{sample}

\item{alpha}{level of confidence}
}
\value{
CI of mean
}
\description{
CI for mean of a sample
}
